+++ module-system/module-implementations/sov-evm/src/evm/test_helpers.rs
@@ -3,17 +3,6 @@ use std::path::PathBuf;
 use ethers_contract::BaseContract;
 use ethers_core::abi::Abi;
 use ethers_core::types::Bytes;
-use revm::primitives::{ExecutionResult, Output};
-
-pub(crate) fn output(result: ExecutionResult) -> bytes::Bytes {
-    match result {
-        ExecutionResult::Success { output, .. } => match output {
-            Output::Call(out) => out,
-            Output::Create(out, _) => out,
-        },
-        _ => panic!("Expected successful ExecutionResult"),
-    }
-}
 
 fn test_data_path() -> PathBuf {
     let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
